requirements.txt
alembic==1.7.7
altgraph==0.17.2
appnope==0.1.2
appscript==1.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.19.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
Eel==0.12.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
ez-setup==0.9
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imap-tools==0.54.0
ipykernel==6.9.1
ipython==8.0.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
keyboard==0.13.5
kiwisolver==1.4.2
lxml==4.8.0
macholib==1.16
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
more-itertools==8.12.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pip-tools==6.6.0
platformdirs==2.5.1
plotly==5.13.0
plotly-express==0.4.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
pygame==2.1.2
Pygments==2.11.2
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.3.0
pyobjc==8.2
pyobjc-core==8.2
pyobjc-framework-Accessibility==8.2
pyobjc-framework-Accounts==8.2
pyobjc-framework-AddressBook==8.2
pyobjc-framework-AdServices==8.2
pyobjc-framework-AdSupport==8.2
pyobjc-framework-AppleScriptKit==8.2
pyobjc-framework-AppleScriptObjC==8.2
pyobjc-framework-ApplicationServices==8.2
pyobjc-framework-AppTrackingTransparency==8.2
pyobjc-framework-AudioVideoBridging==8.2
pyobjc-framework-AuthenticationServices==8.2
pyobjc-framework-AutomaticAssessmentConfiguration==8.2
pyobjc-framework-Automator==8.2
pyobjc-framework-AVFoundation==8.2
pyobjc-framework-AVKit==8.2
pyobjc-framework-BusinessChat==8.2
pyobjc-framework-CalendarStore==8.2
pyobjc-framework-CallKit==8.2
pyobjc-framework-CFNetwork==8.2
pyobjc-framework-ClassKit==8.2
pyobjc-framework-CloudKit==8.2
pyobjc-framework-Cocoa==8.2
pyobjc-framework-Collaboration==8.2
pyobjc-framework-ColorSync==8.2
pyobjc-framework-Contacts==8.2
pyobjc-framework-ContactsUI==8.2
pyobjc-framework-CoreAudio==8.2
pyobjc-framework-CoreAudioKit==8.2
pyobjc-framework-CoreBluetooth==8.2
pyobjc-framework-CoreData==8.2
pyobjc-framework-CoreHaptics==8.2
pyobjc-framework-CoreLocation==8.2
pyobjc-framework-CoreMedia==8.2
pyobjc-framework-CoreMediaIO==8.2
pyobjc-framework-CoreMIDI==8.2
pyobjc-framework-CoreML==8.2
pyobjc-framework-CoreMotion==8.2
pyobjc-framework-CoreServices==8.2
pyobjc-framework-CoreSpotlight==8.2
pyobjc-framework-CoreText==8.2
pyobjc-framework-CoreWLAN==8.2
pyobjc-framework-CryptoTokenKit==8.2
pyobjc-framework-DataDetection==8.2
pyobjc-framework-DeviceCheck==8.2
pyobjc-framework-DictionaryServices==8.2
pyobjc-framework-DiscRecording==8.2
pyobjc-framework-DiscRecordingUI==8.2
pyobjc-framework-DiskArbitration==8.2
pyobjc-framework-DVDPlayback==8.2
pyobjc-framework-EventKit==8.2
pyobjc-framework-ExceptionHandling==8.2
pyobjc-framework-ExecutionPolicy==8.2
pyobjc-framework-ExternalAccessory==8.2
pyobjc-framework-FileProvider==8.2
pyobjc-framework-FileProviderUI==8.2
pyobjc-framework-FinderSync==8.2
pyobjc-framework-FSEvents==8.2
pyobjc-framework-GameCenter==8.2
pyobjc-framework-GameController==8.2
pyobjc-framework-GameKit==8.2
pyobjc-framework-GameplayKit==8.2
pyobjc-framework-ImageCaptureCore==8.2
pyobjc-framework-IMServicePlugIn==8.2
pyobjc-framework-InputMethodKit==8.2
pyobjc-framework-InstallerPlugins==8.2
pyobjc-framework-InstantMessage==8.2
pyobjc-framework-Intents==8.2
pyobjc-framework-IntentsUI==8.2
pyobjc-framework-IOSurface==8.2
pyobjc-framework-iTunesLibrary==8.2
pyobjc-framework-KernelManagement==8.2
pyobjc-framework-LatentSemanticMapping==8.2
pyobjc-framework-LaunchServices==8.2
pyobjc-framework-libdispatch==8.2
pyobjc-framework-LinkPresentation==8.2
pyobjc-framework-LocalAuthentication==8.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.2
pyobjc-framework-MailKit==8.2
pyobjc-framework-MapKit==8.2
pyobjc-framework-MediaAccessibility==8.2
pyobjc-framework-MediaLibrary==8.2
pyobjc-framework-MediaPlayer==8.2
pyobjc-framework-MediaToolbox==8.2
pyobjc-framework-Metal==8.2
pyobjc-framework-MetalKit==8.2
pyobjc-framework-MetalPerformanceShaders==8.2
pyobjc-framework-MetalPerformanceShadersGraph==8.2
pyobjc-framework-MetricKit==8.2
pyobjc-framework-MLCompute==8.2
pyobjc-framework-ModelIO==8.2
pyobjc-framework-MultipeerConnectivity==8.2
pyobjc-framework-NaturalLanguage==8.2
pyobjc-framework-NetFS==8.2
pyobjc-framework-Network==8.2
pyobjc-framework-NetworkExtension==8.2
pyobjc-framework-NotificationCenter==8.2
pyobjc-framework-OpenDirectory==8.2
pyobjc-framework-OSAKit==8.2
pyobjc-framework-OSLog==8.2
pyobjc-framework-PassKit==8.2
pyobjc-framework-PencilKit==8.2
pyobjc-framework-Photos==8.2
pyobjc-framework-PhotosUI==8.2
pyobjc-framework-PreferencePanes==8.2
pyobjc-framework-PushKit==8.2
pyobjc-framework-Quartz==8.2
pyobjc-framework-QuickLookThumbnailing==8.2
pyobjc-framework-ReplayKit==8.2
pyobjc-framework-SafariServices==8.2
pyobjc-framework-SceneKit==8.2
pyobjc-framework-ScreenSaver==8.2
pyobjc-framework-ScreenTime==8.2
pyobjc-framework-ScriptingBridge==8.2
pyobjc-framework-SearchKit==8.2
pyobjc-framework-Security==8.2
pyobjc-framework-SecurityFoundation==8.2
pyobjc-framework-SecurityInterface==8.2
pyobjc-framework-ServiceManagement==8.2
pyobjc-framework-ShazamKit==8.2
pyobjc-framework-Social==8.2
pyobjc-framework-SoundAnalysis==8.2
pyobjc-framework-Speech==8.2
pyobjc-framework-SpriteKit==8.2
pyobjc-framework-StoreKit==8.2
pyobjc-framework-SyncServices==8.2
pyobjc-framework-SystemConfiguration==8.2
pyobjc-framework-SystemExtensions==8.2
pyobjc-framework-UniformTypeIdentifiers==8.2
pyobjc-framework-UserNotifications==8.2
pyobjc-framework-UserNotificationsUI==8.2
pyobjc-framework-VideoSubscriberAccount==8.2
pyobjc-framework-VideoToolbox==8.2
pyobjc-framework-Virtualization==8.2
pyobjc-framework-Vision==8.2
pyobjc-framework-WebKit==8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytils==0.4.1
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
scipy==1.10.0
seaborn==0.12.2
selenium==4.1.5
selenium-elements==0.0.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.35
stack-data==0.2.0
statsmodels==0.13.5
stem==1.8.0
tenacity==8.1.0
tkcalendar==1.6.1
tomli==2.0.1
tools==0.1.9
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
twilio==7.9.0
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.5.4
Werkzeug==2.0.3
whichcraft==0.6.1
wsproto==1.1.0
XlsxWriter==3.0.3
zope.event==4.5.0
zope.interface==5.4.0
